bug-fix: chat parsing error
